More refactoring.  Now uses ticking system to update UI
